set up phan 3
