+++ src/app/(ship-planning)/ke-hoach/CMC/page.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function CMC() {
+  return (
+    <div>
+      CMC
+    </div>
+  )
+}
+
+export default CMC
