+++ src/app/(ship-planning)/ke-hoach/cap-nhat-danh-sach-container-xuat-tau/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function UpdateListContainerExport() {
+  return <div>UpdateListContainerExport</div>;
+}
+
+export default UpdateListContainerExport;
