+++ src/app/(ship-planning)/ke-hoach/danh-sach-container-xuat-tau/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function ListContainerExport() {
+  return <div>ListContainerExport</div>;
+}
+
+export default ListContainerExport
