+++ src/app/(ship-planning)/ke-hoach/gan-cau/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function CraneAllocation() {
+  return <div>CraneAllocation</div>;
+}
+
+export default CraneAllocation;
