+++ src/app/(ship-planning)/ke-hoach/giam-sat-bai/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function YardMonitor() {
+  return <div>YardMonitor</div>;
+}
+
+export default YardMonitor;
