+++ src/app/(ship-planning)/ke-hoach/in-ke-hoach/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function PrintPlan() {
+  return <div>PrintPlan</div>;
+}
+
+export default PrintPlan;
