+++ src/app/(ship-planning)/ke-hoach/ke-hoach-do-container/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function ContainerDischargePlan() {
+  return <div>ContainerDischargePlan</div>;
+}
+
+export default ContainerDischargePlan;
