+++ src/app/(ship-planning)/ke-hoach/ke-hoach-tau/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function ShipPlan() {
+  return <div>ShipPlan</div>;
+}
+
+export default ShipPlan;
