+++ src/app/(ship-planning)/ke-hoach/ke-hoach-xep-container/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function ContainerLoadingPlan() {
+  return <div>ContainerLoadingPlan</div>;
+}
+
+export default ContainerLoadingPlan;
