+++ src/app/(ship-planning)/ke-hoach/out-bound-preplan/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function OutBoundPreplan() {
+  return <div>OutBoundPreplan</div>;
+}
+
+export default OutBoundPreplan;
