+++ src/app/(ship-planning)/ke-hoach/thong-ke-ke-hoach-xep-do/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function HandlingPlanStatistic() {
+  return <div>HandlingPlanStatistic</div>;
+}
+
+export default HandlingPlanStatistic;
